Version 5.0 - New branch: LoadingBy_SQL
Obsoleted the UPDATE part of the program since the new concept is to truncate the table and build a new one from scratch due to the rack areas for server.
Added a function in delete.py to truncate the table.
Added a function in update.py to update the OrderType and FactoryStatus columns that have NULL values.
Changed the method of INSERTING values by uploading in chunks of 1,000.