Version 5.1 - New branch: LoadingBy_SQL_2
Disabled obsolete lines of code, that do not longer apply to the concept of re-building the SQL WIP table from scratch every time the script runs.
Re-modified the items_to_SQL_values() function in utilities.py to build the VALUES part of the query by string multiplication instead of a loop.
Re-modified the logic of INSERT in loading.py to go along with the modification of items_to_SQL_values().