Version 6.0
Fixed a wrong time calculation in model.py. It was missing the 'working time' flag in the dwell time calculation.

Changed the way that FactoryStatus is assigned to each WIP instance. Instead of getting the latest SAP status from the beginning with a join of PH and SAP, now SAP status is extracted separately and all the extracted data is passed as an argument to assign_wip() in transform.py, together with the PH raw data. Then as the cleaning loop is moving for each serial number instance, it gets the SAP status serial number, if it exists, and passes it to the constructor of each ServerHistory or RackHistory object, which is then leveraged by the determine_processAndArea() method to get the correct SAP status for each WIP instance.