Version 6.3
Optimize the process flow to reduce overhead significantly

In model.py replaced the redundant process of finding the day transaction history by using the MAX timestamp as a filter by now using the first row 'iloc[0]' since the data is already sorted in descending order from a previous operation.
Also, removed the working dwell time calculation from here.

In transform.py, the working dwell time calculation is implemented again in the assign_wip() function, but this time, instead of being an element-wise operation on the dataframe, as it was previously, it first separates the transaction and wip_snapshot datetime data into tuples, then the delta_working_hours() function is applied at the tuple level by iterating over all the tuples. All the results from the calculation is stored into a list, and once the calculation operation is complete, it is re-integrated back to the WIP dataframe. This helps reduce overhead significantly.