Version 8.0
Obsoleted the RackHistory class and created a generic one UnitHistory which will serve for both rack and server units. The flag isServer will help identify the type of unit.
In model.py, removed the instances of dt.now() by a datetime variable passed to the constructor, this should reduce some overhead to the loop controlling the UnitHistory instances.