Version 9.1
Modified some bugs. In utilities.py 'df_splitter()' function the partition sizes were set up incorrectly.
Got rid of the dataframe copy() arguments in main.py as well as the manual garbage collector.
Followed pandas advice in model.py to use copy() in this line: location_row = day_ckps_df.iloc[0].copy()
This ensures that location_row is an actual object and not a view of a dataframe. (The recommendation came from this pandas warning: "SettingWithCopyWarning